requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
beautifulsoup4>=4.12.0
mwparserfromhell>=0.6.5
schedule>=1.2.0
python-dotenv>=1.0.0
httpx>=0.25.0
openai>=1.0.0
